Added data validation

+++ api/Controllers/CommentController.cs
@@ -6,6 +6,7 @@ using api.Dtos.Comment;
 using api.Interfaces;
 using api.Mappers;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore.Metadata.Internal;
 
 namespace api.Controllers
 {
@@ -23,15 +24,23 @@ namespace api.Controllers
 
         [HttpGet]
         public async Task<IActionResult> GetAll(){
+
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+
             var comments = await _commentRepo.GetAllAsync();
 
             var commnetDto = comments.Select(s => s.ToCommentDto());
             return Ok(commnetDto);
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id){
 
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var comment = await _commentRepo.GetByIdAsync(id);
 
             if(comment == null){
@@ -41,9 +50,12 @@ namespace api.Controllers
             return  Ok(comment.ToCommentDto());
         }
         
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto){
 
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             if(!await _stockRepo.StockExist(stockId)){
                 return BadRequest("Stock doesn't exist");
             }
@@ -53,8 +65,12 @@ namespace api.Controllers
             return CreatedAtAction(nameof(GetById), new { id = commentModel.Id }, commentModel.ToCommentDto());
         }
 
-        [HttpPut("{id}")]
+        [HttpPut("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id,[FromBody] UpdateCommentDto commentDto){
+
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             if(await _commentRepo.CommentExist(id)){
                 var commentModel = await _commentRepo.UpdateAsync(id, commentDto.ToCommentFromUpdate(id));
                 
@@ -65,8 +81,12 @@ namespace api.Controllers
             return NotFound("Comment not found");
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id :int}")]
         public async Task<IActionResult> Delete([FromRoute] int id){
+
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
 
             var commentModel = await _commentRepo.DeleteAsync(id);
